stream and table initialization separated
